Use docker-ce, not moby for 2.0 (#104)
Moby (Docker version shipped with Fedora) does not work for us,
so let's stick with the proper Docker CE.

This wasn't seen before because our jobs were using the 1.2 image.
We we switched to 2.0 this error showed up.

Also, the previous workaround with the location of some docker tools
is not necessaty anymore.