MAISTRA-2717: Add automation for merging upstream changes
This adds a slightly modified version of the automator.sh
script present in Istio. It has support for merging upstream
repository changes into downstream. If the merge fails, an issue
will be created and maintainers notified.

This also adds a new builder image for the "main" branch that
works for both Istio and Envoy/Proxy repositories.

Finally this adds a periodic job that runs once a day triggering
the merge of envoy/envoy@main into maistra/envoy@main.