Fix the daily build
By setting the `BUILD_WITH_CONTAINER` variable to "0". Without it, the istio
build was happening within a nested container, and, as the variable
`DOCKER_CONFIG` is not being forwarded to this nested container, the
push to quay.io fails due the missing credentials.

We don't need this nested container as the build already happens in our
builder image, as it is with the pre-submit jobs.