Fix daily build for 2.3 (#286)
By making the docker config directory writable. `docker buildx` creates
a `buildx` subdirectory in it. If the volume is read only, the job
fails.

While on it, also change the name of the volume mounts so it doesn't
clash with other mounts. Change from `creds` to `docker-config` which is
more accurate.