Do not rely on OpenSSL from distro (#407)
Because it's rolling new versions and we have no control over it.
Instead, install the open source version of OpenSSL.

By setting the `OPENSSL_ROOT_DIR` variable, cmake will be able to find
our OpenSSL and build/link against it. Also we populate the
`/etc/ld.so.conf.d/openssl.conf` file so that the correct libraries can
be loaded at runtime.