Install go 1.22 along with go 1.20 (#413)
Instead of replacing go 1.20 with 1.22, install 1.22 in a separate
directory so that Envoy/Proxy can use it by adjusting the `$PATH`.

Istio-operatori 2.5 has a hard requirement on go 1.20, thus this PR
brings the peace back.

`gocovmerge` was having issues to build, so I disabled it, I don't think
we use it anyway.